backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic-settings==2.4.0
httpx==0.27.0
python-multipart==0.0.9
sqlmodel==0.0.22
sqladmin==0.16.1
jinja2==3.1.4
